Create Person and mission, everything s working

+++ src/main/java/com/blog/soat/graphql/repository/PersonRepository.java
@@ -1,7 +1,7 @@
 package com.blog.soat.graphql.repository;
 
-import com.blog.soat.graphql.model.Person;
+import com.blog.soat.graphql.entities.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person, Integer> {
+public interface PersonRepository extends CrudRepository<Person, Long> {
 }
